--- app/assets/css/main.css
@@ -1,6 +1,0 @@
-/* Normalize.css */
-@import '@unocss/reset/normalize.css';
-
-/* sanitize.css */
-@import '@unocss/reset/sanitize/sanitize.css';
-@import '@unocss/reset/sanitize/assets.css';
